+++ src/test/java/io/supertokens/test/emailpassword/PasswordHashingTest.java
@@ -25,6 +25,7 @@ import io.supertokens.emailpassword.exceptions.WrongCredentialsException;
 import io.supertokens.inmemorydb.Start;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.pluginInterface.emailpassword.UserInfo;
+import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.storageLayer.StorageLayer;
 import io.supertokens.test.TestingProcessManager;
 import io.supertokens.test.Utils;
@@ -38,6 +39,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.Assert.*;
 
@@ -58,8 +60,6 @@ public class PasswordHashingTest {
     @Test
     public void hashAndVerifyWithBcrypt() throws Exception {
         String[] args = { "../" };
-
-        Utils.setValueInConfig("password_hashing_alg", "BCRYPT");
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
@@ -123,7 +123,6 @@ public class PasswordHashingTest {
         String hash = "";
 
         {
-            Utils.setValueInConfig("password_hashing_alg", "BCRYPT");
 
             TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
@@ -132,7 +131,7 @@ public class PasswordHashingTest {
                 return;
             }
 
-            hash = PasswordHashing.createHashWithSalt(process.getProcess(), "somePassword");
+            hash = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePassword");
 
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_BCRYPT));
 
@@ -147,7 +146,7 @@ public class PasswordHashingTest {
             assert (Config.getConfig(process.getProcess())
                     .getPasswordHashingAlg() == CoreConfig.PASSWORD_HASHING_ALG.ARGON2);
 
-            assert (PasswordHashing.verifyPasswordWithHash(process.getProcess(), "somePassword", hash));
+            assert (PasswordHashing.getInstance(process.getProcess()).verifyPasswordWithHash("somePassword", hash));
 
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_VERIFY_BCRYPT));
 
@@ -171,7 +170,7 @@ public class PasswordHashingTest {
                 return;
             }
 
-            hash = PasswordHashing.createHashWithSalt(process.getProcess(), "somePassword");
+            hash = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePassword");
 
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_ARGON));
 
@@ -186,7 +185,7 @@ public class PasswordHashingTest {
             assert (Config.getConfig(process.getProcess())
                     .getPasswordHashingAlg() == CoreConfig.PASSWORD_HASHING_ALG.BCRYPT);
 
-            PasswordHashing.verifyPasswordWithHash(process.getProcess(), "somePassword", hash);
+            PasswordHashing.getInstance(process.getProcess()).verifyPasswordWithHash("somePassword", hash);
 
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_VERIFY_ARGON));
 
@@ -196,7 +195,7 @@ public class PasswordHashingTest {
     }
 
     @Test
-    public void defaultConfigIsArgon() throws Exception {
+    public void defaultConfigs() throws Exception {
         String[] args = { "../" };
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -208,13 +207,65 @@ public class PasswordHashingTest {
 
         CoreConfig config = Config.getConfig(process.getProcess());
 
-        assert (config.getPasswordHashingAlg() == CoreConfig.PASSWORD_HASHING_ALG.ARGON2);
+        assert (config.getPasswordHashingAlg() == CoreConfig.PASSWORD_HASHING_ALG.BCRYPT);
         assert (config.getArgon2Iterations() == 3);
         assert (config.getArgon2MemoryBytes() == 65536);
-        assert (config.getArgon2Parallelism() == 1);
+        assert (config.getArgon2Parallelism() == 4);
+        assert (config.getArgon2HashingPoolSize() == 10);
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void invalidConfigArgonButUsingBcryptShouldAllowStartingServer() throws Exception {
+        {
+            String[] args = { "../" };
+            Utils.setValueInConfig("argon2_memory_bytes", "-1");
+            Utils.setValueInConfig("argon2_parallelism", "-1");
+            Utils.setValueInConfig("argon2_iterations", "-1");
+            Utils.setValueInConfig("argon2_hashing_pool_size", "-1");
+
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+            EmailPassword.signUp(process.getProcess(), "t@a.com", "somePassword");
+            EmailPassword.signIn(process.getProcess(), "t@a.com", "somePassword");
+
+            process.kill();
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
+    }
+
+    @Test
+    public void lowercaseConfig() throws Exception {
+        {
+            String[] args = { "../" };
+            Utils.setValueInConfig("password_hashing_alg", "argon2");
+
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+            assert (Config.getConfig(process.getProcess())
+                    .getPasswordHashingAlg() == CoreConfig.PASSWORD_HASHING_ALG.ARGON2);
+
+            process.kill();
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
+
+        Utils.reset();
+
+        {
+            String[] args = { "../" };
+            Utils.setValueInConfig("password_hashing_alg", "bcrypt");
+
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+            assert (Config.getConfig(process.getProcess())
+                    .getPasswordHashingAlg() == CoreConfig.PASSWORD_HASHING_ALG.BCRYPT);
+
+            process.kill();
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
     }
 
     @Test
@@ -237,6 +288,7 @@ public class PasswordHashingTest {
         {
             String[] args = { "../" };
             Utils.setValueInConfig("argon2_iterations", "-1");
+            Utils.setValueInConfig("password_hashing_alg", "argon2");
 
             TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
             ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
@@ -252,6 +304,7 @@ public class PasswordHashingTest {
         {
             String[] args = { "../" };
             Utils.setValueInConfig("argon2_parallelism", "-1");
+            Utils.setValueInConfig("password_hashing_alg", "ARGON2");
 
             TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
             ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
@@ -267,11 +320,44 @@ public class PasswordHashingTest {
         {
             String[] args = { "../" };
             Utils.setValueInConfig("argon2_memory_bytes", "-1");
+            Utils.setValueInConfig("password_hashing_alg", "ARGON2");
 
             TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
             ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
             assertNotNull(e);
             assertEquals(e.exception.getMessage(), "'argon2_memory_bytes' must be >= 1");
+
+            process.kill();
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
+
+        Utils.reset();
+
+        {
+            String[] args = { "../" };
+            Utils.setValueInConfig("argon2_hashing_pool_size", "-1");
+            Utils.setValueInConfig("password_hashing_alg", "argon2");
+
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
+            assertNotNull(e);
+            assertEquals(e.exception.getMessage(), "'argon2_hashing_pool_size' must be >= 1");
+
+            process.kill();
+            assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
+
+        Utils.reset();
+
+        {
+            String[] args = { "../" };
+            Utils.setValueInConfig("argon2_hashing_pool_size", "100");
+            Utils.setValueInConfig("password_hashing_alg", "ARGON2");
+
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            ProcessState.EventAndException e = process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.INIT_FAILURE);
+            assertNotNull(e);
+            assertEquals(e.exception.getMessage(), "'argon2_hashing_pool_size' must be <= 'max_server_pool_size'");
 
             process.kill();
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
@@ -282,6 +368,7 @@ public class PasswordHashingTest {
     public void hashAndVerufyArgon2HashWithDifferentConfigs() throws Exception {
         String[] args = { "../" };
         String hash = "";
+        Utils.setValueInConfig("password_hashing_alg", "ARGON2");
 
         {
             TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -291,7 +378,7 @@ public class PasswordHashingTest {
                 return;
             }
 
-            hash = PasswordHashing.createHashWithSalt(process.getProcess(), "somePassword");
+            hash = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePassword");
 
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_ARGON));
 
@@ -303,17 +390,19 @@ public class PasswordHashingTest {
             Utils.setValueInConfig("argon2_memory_bytes", "100");
             Utils.setValueInConfig("argon2_parallelism", "2");
             Utils.setValueInConfig("argon2_iterations", "10");
+            Utils.setValueInConfig("argon2_hashing_pool_size", "5");
             TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
-            assert (PasswordHashing.verifyPasswordWithHash(process.getProcess(), "somePassword", hash));
+            assert (PasswordHashing.getInstance(process.getProcess()).verifyPasswordWithHash("somePassword", hash));
 
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_VERIFY_ARGON));
 
-            String newHash = PasswordHashing.createHashWithSalt(process.getProcess(), "somePassword");
+            String newHash = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePassword");
             assert (newHash.contains("m=100"));
             assert (newHash.contains("p=2"));
             assert (newHash.contains("t=10"));
+            assert (Config.getConfig(process.getProcess()).getArgon2HashingPoolSize() == 5);
 
             process.kill();
             assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
@@ -323,8 +412,6 @@ public class PasswordHashingTest {
     @Test
     public void hashAndVerifyWithBcryptChangeToArgonPasswordWithResetFlow() throws Exception {
         String[] args = { "../" };
-
-        Utils.setValueInConfig("password_hashing_alg", "BCRYPT");
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
@@ -356,6 +443,7 @@ public class PasswordHashingTest {
     public void hashAndVerifyWithArgonChangeToBcryptPasswordWithResetFlow() throws Exception {
         String[] args = { "../" };
 
+        Utils.setValueInConfig("password_hashing_alg", "ARGON2");
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
@@ -386,8 +474,6 @@ public class PasswordHashingTest {
     public void hashAndVerifyWithBcryptChangeToArgonChangePassword() throws Exception {
         String[] args = { "../" };
 
-        Utils.setValueInConfig("password_hashing_alg", "BCRYPT");
-
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
@@ -417,6 +503,7 @@ public class PasswordHashingTest {
     public void hashAndVerifyWithArgonChangeToBcryptChangePassword() throws Exception {
         String[] args = { "../" };
 
+        Utils.setValueInConfig("password_hashing_alg", "ARGON2");
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
@@ -443,7 +530,29 @@ public class PasswordHashingTest {
     }
 
     @Test
-    public void differentPasswordHashGenerated() throws Exception {
+    public void differentPasswordHashGeneratedArgon() throws Exception {
+        String[] args = { "../" };
+
+        Utils.setValueInConfig("password_hashing_alg", "ARGON2");
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
+            return;
+        }
+
+        String hash = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePass");
+        String hash2 = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePass");
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_ARGON));
+
+        assert (!hash.equals(hash2));
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void differentPasswordHashGeneratedBcrypt() throws Exception {
         String[] args = { "../" };
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -453,9 +562,9 @@ public class PasswordHashingTest {
             return;
         }
 
-        String hash = PasswordHashing.createHashWithSalt(process.getProcess(), "somePass");
-        String hash2 = PasswordHashing.createHashWithSalt(process.getProcess(), "somePass");
-        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_ARGON));
+        String hash = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePass");
+        String hash2 = PasswordHashing.getInstance(process.getProcess()).createHashWithSalt("somePass");
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_BCRYPT));
 
         assert (!hash.equals(hash2));
 
@@ -467,6 +576,7 @@ public class PasswordHashingTest {
     public void parallelSignUpSignIn() throws Exception {
         String[] args = { "../" };
 
+        Utils.setValueInConfig("password_hashing_alg", "ARGON2");
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
@@ -477,17 +587,37 @@ public class PasswordHashingTest {
             return;
         }
 
-        AtomicBoolean success = new AtomicBoolean(true);
+        AtomicInteger counter = new AtomicInteger(0);
+
+        assert (PasswordHashing.getInstance(process.getProcess()).getBlockedQueueSize() == 0);
+        AtomicBoolean reachedQueueMaxSize = new AtomicBoolean(false);
 
         ExecutorService ex = Executors.newFixedThreadPool(1000);
-        for (int i = 0; i < 50; i++) {
-            int y = i;
+        int numberOfThreads = 500;
+        for (int i = 0; i < numberOfThreads; i++) {
+            int finalI = i;
             ex.execute(() -> {
-                try {
-                    EmailPassword.signUp(process.getProcess(), "test@example.com" + y, "somePassword" + y);
-                    EmailPassword.signIn(process.getProcess(), "test@example.com" + y, "somePassword" + y);
-                } catch (Exception e) {
-                    success.set(false);
+                int localCounter = 0;
+                while (true) {
+                    String uniqueEmail = "test@example.com" + finalI + "" + localCounter;
+                    localCounter++;
+                    try {
+                        EmailPassword.signUp(process.getProcess(), uniqueEmail, "somePassword" + finalI);
+                        EmailPassword.signIn(process.getProcess(), uniqueEmail, "somePassword" + finalI);
+                        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.PASSWORD_HASH_ARGON));
+                        int queueSize = PasswordHashing.getInstance(process.getProcess()).getBlockedQueueSize();
+                        int maxQueueSize = Config.getConfig(process.getProcess()).getArgon2HashingPoolSize();
+                        assert (queueSize <= maxQueueSize);
+                        if (queueSize == maxQueueSize || queueSize + 1 == maxQueueSize) {
+                            reachedQueueMaxSize.set(true);
+                        }
+                        counter.incrementAndGet();
+                        break;
+                    } catch (StorageQueryException e) {
+                        // we try again as this may happen cause of connection timeout in db layer.
+                    } catch (Exception ignored) {
+                        break;
+                    }
                 }
             });
         }
@@ -496,7 +626,8 @@ public class PasswordHashingTest {
 
         ex.awaitTermination(2, TimeUnit.MINUTES);
 
-        assert (success.get());
+        assert (counter.get() == numberOfThreads);
+        assert (reachedQueueMaxSize.get());
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));